requirements.txt
Flask==2.0.1
opencv-python-headless==4.5.3.56
numpy==1.21.2
scipy==1.7.1
Pillow==8.3.2
gunicorn==20.1.0